RL_researcher_writer_ymaxing/research_agent_local/mcp_client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "The Nova MCP client"
readme = "README.md"
requires-python = "==3.12.11"
dependencies = [
    "rich>=14.0.0",
    "opik>=1.8.42",
    "langgraph>=0.5.3",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.1.8",
    "langchain-openai>=0.3.28",
    "langchain-perplexity>=0.1.1",
    "langchain-tavily>= 0.2.0",
    "langchain-mcp-adapters>=0.1.9",
    "pydantic-settings>=2.10.0",
    "google-genai>=1.20.0",
    "fastmcp>=0.9.0",
    "mcp-agent>=0.2.0",
    "python-dotenv>=1.0.0",
    "firecrawl>=4.3.6",
    "gitingest>=0.1.5",
    "httpx>=0.28.1",
    "mcp>=1.11.0",
    "nbformat>=5.10.4",
    "psutil>=7.0.0",
]